music_app/requirements.txt
cassandra-driver>=3.25.0
python-dotenv>=0.19.0